@@ -1,6 +1,6 @@
-// sequence-api v0.4.0 c3eb9010746703c095291691cd8f08dc04999062
+// sequence-api v0.4.0 1ac0fac314dfe117706d15f13a63d49bdcec048e
 // --
-// Code generated by webrpc-gen@v0.22.0 with golang generator. DO NOT EDIT.
+// Code generated by webrpc-gen@v0.23.2 with golang generator. DO NOT EDIT.
 //
 // webrpc-gen -schema=api.ridl -target=golang -pkg=api -client -out=./clients/api.gen.go
 package api
@@ -22,7 +22,7 @@ import (
 
 const WebrpcHeader = "Webrpc"
 
-const WebrpcHeaderValue = "webrpc@v0.22.0;gen-golang@v0.17.0;sequence-api@v0.4.0"
+const WebrpcHeaderValue = "webrpc@v0.23.2;gen-golang@v0.17.0;sequence-api@v0.4.0"
 
 // WebRPC description and code-gen version
 func WebRPCVersion() string {
@@ -36,7 +36,7 @@ func WebRPCSchemaVersion() string {
 
 // Schema hash generated from your RIDL schema
 func WebRPCSchemaHash() string {
-	return "c3eb9010746703c095291691cd8f08dc04999062"
+	return "1ac0fac314dfe117706d15f13a63d49bdcec048e"
 }
 
 type WebrpcGenVersions struct {
@@ -277,6 +277,51 @@ func (x *TokenType) Is(values ...TokenType) bool {
 	return false
 }
 
+type TransakBuySell uint8
+
+const (
+	TransakBuySell_UNKNOWN TransakBuySell = 0
+	TransakBuySell_BUY     TransakBuySell = 1
+	TransakBuySell_SELL    TransakBuySell = 2
+)
+
+var TransakBuySell_name = map[uint8]string{
+	0: "UNKNOWN",
+	1: "BUY",
+	2: "SELL",
+}
+
+var TransakBuySell_value = map[string]uint8{
+	"UNKNOWN": 0,
+	"BUY":     1,
+	"SELL":    2,
+}
+
+func (x TransakBuySell) String() string {
+	return TransakBuySell_name[uint8(x)]
+}
+
+func (x TransakBuySell) MarshalText() ([]byte, error) {
+	return []byte(TransakBuySell_name[uint8(x)]), nil
+}
+
+func (x *TransakBuySell) UnmarshalText(b []byte) error {
+	*x = TransakBuySell(TransakBuySell_value[string(b)])
+	return nil
+}
+
+func (x *TransakBuySell) Is(values ...TransakBuySell) bool {
+	if x == nil {
+		return false
+	}
+	for _, v := range values {
+		if *x == v {
+			return true
+		}
+	}
+	return false
+}
+
 type Version struct {
 	WebrpcVersion string `json:"webrpcVersion"`
 	SchemaVersion string `json:"schemaVersion"`
@@ -400,14 +445,29 @@ type TupleComponent struct {
 	Value interface{} `json:"value"`
 }
 
+type IntentPrecondition struct {
+	Type    string          `json:"type"`
+	ChainID prototyp.BigInt `json:"chainId"`
+	Data    interface{}     `json:"data"`
+}
+
+type IntentSolution struct {
+	Transactions []*Transactions `json:"transactions"`
+}
+
+type Transactions struct {
+	ChainID       prototyp.BigInt       `json:"chainId"`
+	Transactions  []*Transaction        `json:"transactions"`
+	Preconditions []*IntentPrecondition `json:"preconditions,omitempty"`
+}
+
 type Transaction struct {
 	DelegateCall  bool            `json:"delegateCall"`
 	RevertOnError bool            `json:"revertOnError"`
 	GasLimit      prototyp.BigInt `json:"gasLimit"`
-	Target        string          `json:"target"`
+	Target        prototyp.Hash   `json:"target"`
 	Value         prototyp.BigInt `json:"value"`
-	Data          string          `json:"data"`
-	Call          *ContractCall   `json:"call"`
+	Data          prototyp.Hash   `json:"data"`
 }
 
 type UserStorage struct {
@@ -647,6 +707,29 @@ type SardinePaymentOption struct {
 	ProcessingTime string   `json:"processingTime"`
 }
 
+type SwapPermit2Price struct {
+	CurrencyAddress string          `json:"currencyAddress"`
+	CurrencyBalance prototyp.BigInt `json:"currencyBalance"`
+	Price           prototyp.BigInt `json:"price"`
+	// Guaranteed price for the swap (incl slippage).
+	MaxPrice         prototyp.BigInt `json:"maxPrice"`
+	TransactionValue prototyp.BigInt `json:"transactionValue"`
+}
+
+type SwapPermit2Quote struct {
+	CurrencyAddress string          `json:"currencyAddress"`
+	CurrencyBalance prototyp.BigInt `json:"currencyBalance"`
+	Price           prototyp.BigInt `json:"price"`
+	// Guaranteed price for the swap (incl slippage).
+	MaxPrice prototyp.BigInt `json:"maxPrice"`
+	// to must be approved to access maxPrice of currencyAddress for the sender
+	To               string          `json:"to"`
+	TransactionData  string          `json:"transactionData"`
+	TransactionValue prototyp.BigInt `json:"transactionValue"`
+	// Supplied when includeApprove is true
+	ApproveData string `json:"approveData"`
+}
+
 type SwapPrice struct {
 	CurrencyAddress string          `json:"currencyAddress"`
 	CurrencyBalance prototyp.BigInt `json:"currencyBalance"`
@@ -729,6 +812,174 @@ type PaymentResponse struct {
 
 type AdoptedChildWallet struct {
 	Address string `json:"address"`
+}
+
+type Pack struct {
+	ChainID         uint64         `json:"chainId" db:"chain_id"`
+	ProjectID       uint64         `json:"projectId,omitempty" db:"project_id"`
+	ContractAddress prototyp.Hash  `json:"contractAddress" db:"contract_address"`
+	Content         []*PackContent `json:"content" db:"content"`
+	CreatedAt       *time.Time     `json:"createdAt,omitempty" db:"created_at,omitempty"`
+}
+
+type PackContent struct {
+	TokenAddresses []string   `json:"tokenAddresses"`
+	TokenIds       [][]string `json:"tokenIds"`
+	Amounts        [][]string `json:"amounts"`
+}
+
+type TransakCountry struct {
+	Alpha2             string            `json:"alpha2"`
+	Alpha3             string            `json:"alpha3"`
+	IsAllowed          bool              `json:"isAllowed"`
+	IsLightKycAllowed  bool              `json:"isLightKycAllowed"`
+	Name               string            `json:"name"`
+	CurrencyCode       string            `json:"currencyCode"`
+	SupportedDocuments []string          `json:"supportedDocuments"`
+	Partners           []*TransakPartner `json:"partners"`
+	States             []*TransakState   `json:"states"`
+}
+
+type TransakPartner struct {
+	Name          string `json:"name"`
+	IsCardPayment bool   `json:"isCardPayment"`
+	CurrencyCode  string `json:"currencyCode"`
+}
+
+type TransakState struct {
+	Code      string `json:"code"`
+	Name      string `json:"name"`
+	IsAllowed bool   `json:"isAllowed"`
+}
+
+type TransakCryptoCurrency struct {
+	ID                        string                        `json:"id"`
+	CoinID                    string                        `json:"coinID"`
+	Address                   string                        `json:"address"`
+	AddressAdditionalData     interface{}                   `json:"addressAdditionalData"`
+	CreatedAt                 time.Time                     `json:"createdAt"`
+	Decimals                  int                           `json:"decimals"`
+	Image                     *TransakCryptoCurrencyImage   `json:"image"`
+	IsAllowed                 bool                          `json:"isAllowed"`
+	IsPopular                 bool                          `json:"isPopular"`
+	IsStable                  bool                          `json:"isStable"`
+	Name                      string                        `json:"name"`
+	RoundOff                  int                           `json:"roundOff"`
+	Symbol                    string                        `json:"symbol"`
+	IsIgnorePriceVerification bool                          `json:"isIgnorePriceVerification"`
+	ImageBk                   *TransakCryptoCurrencyImage   `json:"imageBk"`
+	KycCountriesNotSupported  []string                      `json:"kycCountriesNotSupported"`
+	Network                   *TransakCryptoCurrencyNetwork `json:"network"`
+	UniqueID                  string                        `json:"uniqueID"`
+	TokenType                 string                        `json:"tokenType"`
+	TokenIdentifier           string                        `json:"tokenIdentifier"`
+	IsPayInAllowed            bool                          `json:"isPayInAllowed"`
+	IsSuspended               bool                          `json:"isSuspended"`
+}
+
+type TransakCryptoCurrencyImage struct {
+	Large string `json:"large"`
+	Small string `json:"small"`
+	Thumb string `json:"thumb"`
+}
+
+type TransakCryptoCurrencyNetwork struct {
+	Name                       string        `json:"name"`
+	FiatCurrenciesNotSupported []interface{} `json:"fiatCurrenciesNotSupported"`
+	ChainID                    string        `json:"chainID"`
+}
+
+type TransakCryptoCurrencyNetworkFiatNotSupported struct {
+	FiatCurrency  string `json:"fiatCurrency"`
+	PaymentMethod string `json:"paymentMethod"`
+}
+
+type TransakFiatCurrency struct {
+	Symbol               string                              `json:"symbol"`
+	SupportingCountries  []string                            `json:"supportingCountries"`
+	LogoSymbol           string                              `json:"logoSymbol"`
+	Name                 string                              `json:"name"`
+	PaymentOptions       []*TransakFiatCurrencyPaymentOption `json:"paymentOptions"`
+	IsPopular            bool                                `json:"isPopular"`
+	IsAllowed            bool                                `json:"isAllowed"`
+	RoundOff             int                                 `json:"roundOff"`
+	IsPayOutAllowed      bool                                `json:"isPayOutAllowed"`
+	DefaultCountryForNFT string                              `json:"defaultCountryForNFT"`
+	Icon                 string                              `json:"icon"`
+	DisplayMessage       string                              `json:"displayMessage"`
+}
+
+type TransakFiatCurrencyPaymentOption struct {
+	Name                      string   `json:"name"`
+	ID                        string   `json:"id"`
+	IsNftAllowed              bool     `json:"isNftAllowed"`
+	IsNonCustodial            bool     `json:"isNonCustodial"`
+	ProcessingTime            string   `json:"processingTime"`
+	DisplayText               bool     `json:"displayText"`
+	Icon                      string   `json:"icon"`
+	LimitCurrency             string   `json:"limitCurrency"`
+	IsActive                  bool     `json:"isActive"`
+	Provider                  string   `json:"provider"`
+	MaxAmount                 int      `json:"maxAmount"`
+	MinAmount                 int      `json:"minAmount"`
+	DefaultAmount             int      `json:"defaultAmount"`
+	IsConverted               bool     `json:"isConverted"`
+	VisaPayoutCountries       []string `json:"visaPayoutCountries"`
+	MastercardPayoutCountries []string `json:"mastercardPayoutCountries"`
+	IsPayOutAllowed           bool     `json:"isPayOutAllowed"`
+	MinAmountForPayOut        int      `json:"minAmountForPayOut"`
+	MaxAmountForPayOut        int      `json:"maxAmountForPayOut"`
+	DefaultAmountForPayOut    int      `json:"defaultAmountForPayOut"`
+}
+
+type TransakPrice struct {
+	QuoteID                 string                      `json:"quoteID"`
+	ConversionPrice         float64                     `json:"conversionPrice"`
+	MarketConversionPrice   float64                     `json:"marketConversionPrice"`
+	Slippage                int                         `json:"slippage"`
+	FiatCurrency            string                      `json:"fiatCurrency"`
+	CryptoCurrency          string                      `json:"cryptoCurrency"`
+	PaymentMethod           string                      `json:"paymentMethod"`
+	FiatAmount              float64                     `json:"fiatAmount"`
+	CryptoAmount            float64                     `json:"cryptoAmount"`
+	IsBuyOrSell             string                      `json:"isBuyOrSell"`
+	Network                 string                      `json:"network"`
+	FeeDecimal              float64                     `json:"feeDecimal"`
+	TotalFee                float64                     `json:"totalFee"`
+	FeeBreakdown            []*TransakPriceFeeBreakdown `json:"feeBreakdown"`
+	Nonce                   int                         `json:"nonce"`
+	CryptoLiquidityProvider string                      `json:"cryptoLiquidityProvider"`
+	Notes                   []interface{}               `json:"notes"`
+}
+
+type TransakPriceFeeBreakdown struct {
+	Name  string   `json:"Name"`
+	Value float64  `json:"Value"`
+	ID    string   `json:"ID"`
+	Ids   []string `json:"Ids"`
+}
+
+type TransakGetPriceParams struct {
+	// required Defaults to GBP
+	FiatCurrency string `json:"fiatCurrency"`
+	// required Defaults to ETH
+	CryptoCurrency string `json:"cryptoCurrency"`
+	// required Defaults to BUY
+	IsBuyOrSell TransakBuySell `json:"isBuyOrSell"`
+	// required Defaults to ethereum
+	Network string `json:"network"`
+	// required Defaults to credit_debit_card Payment method. You can get the payment method id from getFiatCurrencies endpoint
+	PaymentMethod string `json:"paymentMethod"`
+	// Defaults to 100
+	FiatAmount   float64 `json:"fiatAmount"`
+	CryptoAmount float64 `json:"cryptoAmount"`
+	// ISO 3166 Alpha-2 Country Code
+	QuoteCountryCode string `json:"quoteCountryCode"`
+}
+
+type TransakChain struct {
+	Name    string `json:"name"`
+	ChainID uint64 `json:"chainId"`
 }
 
 var methods = map[string]method{
@@ -937,6 +1188,31 @@ var methods = map[string]method{
 		Service:     "API",
 		Annotations: map[string]string{},
 	},
+	"/rpc/API/TransakGetCountries": {
+		Name:        "TransakGetCountries",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
+	"/rpc/API/TransakGetCryptoCurrencies": {
+		Name:        "TransakGetCryptoCurrencies",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
+	"/rpc/API/TransakGetFiatCurrencies": {
+		Name:        "TransakGetFiatCurrencies",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
+	"/rpc/API/TransakGetPrice": {
+		Name:        "TransakGetPrice",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
+	"/rpc/API/TransakGetSupportedNFTCheckoutChains": {
+		Name:        "TransakGetSupportedNFTCheckoutChains",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
 	"/rpc/API/GetCoinPrices": {
 		Name:        "GetCoinPrices",
 		Service:     "API",
@@ -1022,6 +1298,21 @@ var methods = map[string]method{
 		Service:     "API",
 		Annotations: map[string]string{},
 	},
+	"/rpc/API/GetSwapPermit2Price": {
+		Name:        "GetSwapPermit2Price",
+		Service:     "API",
+		Annotations: map[string]string{"public": ""},
+	},
+	"/rpc/API/GetSwapPermit2Prices": {
+		Name:        "GetSwapPermit2Prices",
+		Service:     "API",
+		Annotations: map[string]string{"public": ""},
+	},
+	"/rpc/API/GetSwapPermit2Quote": {
+		Name:        "GetSwapPermit2Quote",
+		Service:     "API",
+		Annotations: map[string]string{"public": ""},
+	},
 	"/rpc/API/GetSwapPrice": {
 		Name:        "GetSwapPrice",
 		Service:     "API",
@@ -1036,6 +1327,16 @@ var methods = map[string]method{
 		Name:        "GetSwapQuote",
 		Service:     "API",
 		Annotations: map[string]string{"public": ""},
+	},
+	"/rpc/API/GetSwapQuoteV2": {
+		Name:        "GetSwapQuoteV2",
+		Service:     "API",
+		Annotations: map[string]string{"public": ""},
+	},
+	"/rpc/API/IntentQuery": {
+		Name:        "IntentQuery",
+		Service:     "API",
+		Annotations: map[string]string{},
 	},
 	"/rpc/API/ListCurrencyGroups": {
 		Name:        "ListCurrencyGroups",
@@ -1074,6 +1375,26 @@ var methods = map[string]method{
 	},
 	"/rpc/API/ListOffchainPayments": {
 		Name:        "ListOffchainPayments",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
+	"/rpc/API/SavePack": {
+		Name:        "SavePack",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
+	"/rpc/API/GetPack": {
+		Name:        "GetPack",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
+	"/rpc/API/DeletePack": {
+		Name:        "DeletePack",
+		Service:     "API",
+		Annotations: map[string]string{},
+	},
+	"/rpc/API/UpdatePackContent": {
+		Name:        "UpdatePackContent",
 		Service:     "API",
 		Annotations: map[string]string{},
 	},
@@ -1131,6 +1452,11 @@ var WebRPCServices = map[string][]string{
 		"GetSardineClientToken",
 		"GetSardineNFTCheckoutToken",
 		"GetSardineNFTCheckoutOrderStatus",
+		"TransakGetCountries",
+		"TransakGetCryptoCurrencies",
+		"TransakGetFiatCurrencies",
+		"TransakGetPrice",
+		"TransakGetSupportedNFTCheckoutChains",
 		"GetCoinPrices",
 		"GetCollectiblePrices",
 		"GetExchangeRate",
@@ -1148,9 +1474,14 @@ var WebRPCServices = map[string][]string{
 		"GenerateWaaSVerificationURL",
 		"ValidateWaaSVerificationNonce",
 		"ListAdoptedWallets",
+		"GetSwapPermit2Price",
+		"GetSwapPermit2Prices",
+		"GetSwapPermit2Quote",
 		"GetSwapPrice",
 		"GetSwapPrices",
 		"GetSwapQuote",
+		"GetSwapQuoteV2",
+		"IntentQuery",
 		"ListCurrencyGroups",
 		"AddOffchainInventory",
 		"GetOffchainInventory",
@@ -1159,6 +1490,10 @@ var WebRPCServices = map[string][]string{
 		"DeleteOffchainInventory",
 		"RequestOffchainPayment",
 		"ListOffchainPayments",
+		"SavePack",
+		"GetPack",
+		"DeletePack",
+		"UpdatePackContent",
 	},
 }
 
@@ -1233,6 +1568,11 @@ type API interface {
 	GetSardineNFTCheckoutToken(ctx context.Context, params *SardineNFTCheckoutParams) (*SardineNFTCheckout, error)
 	// Deprecated. Use SardineGetNFTCheckoutOrderStatus() instead.
 	GetSardineNFTCheckoutOrderStatus(ctx context.Context, orderId string) (*SardineOrder, error)
+	TransakGetCountries(ctx context.Context) ([]*TransakCountry, error)
+	TransakGetCryptoCurrencies(ctx context.Context) ([]*TransakCryptoCurrency, error)
+	TransakGetFiatCurrencies(ctx context.Context) ([]*TransakFiatCurrency, error)
+	TransakGetPrice(ctx context.Context, params *TransakGetPriceParams) (*TransakPrice, error)
+	TransakGetSupportedNFTCheckoutChains(ctx context.Context) ([]*TransakChain, error)
 	//
 	// Price Feed
 	//
@@ -1280,9 +1620,17 @@ type API interface {
 	//
 	// Currency abstraction
 	//
+	GetSwapPermit2Price(ctx context.Context, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) (*SwapPermit2Price, error)
+	GetSwapPermit2Prices(ctx context.Context, userAddress string, buyCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) ([]*SwapPermit2Price, error)
+	GetSwapPermit2Quote(ctx context.Context, userAddress string, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, includeApprove bool, slippagePercentage *uint64) (*SwapPermit2Quote, error)
 	GetSwapPrice(ctx context.Context, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) (*SwapPrice, error)
 	GetSwapPrices(ctx context.Context, userAddress string, buyCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) ([]*SwapPrice, error)
 	GetSwapQuote(ctx context.Context, userAddress string, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, includeApprove bool, slippagePercentage *uint64) (*SwapQuote, error)
+	GetSwapQuoteV2(ctx context.Context, userAddress string, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, includeApprove bool, slippagePercentage *uint64) (*SwapQuote, error)
+	//
+	// Chain abstraction
+	//
+	IntentQuery(ctx context.Context, wallet string, preconditions []*IntentPrecondition) ([]*IntentSolution, error)
 	//
 	// Inventory, payments and management
 	//
@@ -1294,6 +1642,13 @@ type API interface {
 	DeleteOffchainInventory(ctx context.Context, inventoryId uint64) (bool, error)
 	RequestOffchainPayment(ctx context.Context, inventoryId uint64, recipient string, chainId *uint64, tokenAddress *string) (*PaymentResponse, error)
 	ListOffchainPayments(ctx context.Context, inventoryId uint64, page *Page) (*Page, []*OffchainPayment, error)
+	//
+	// Packs
+	//
+	SavePack(ctx context.Context, pack *Pack) (string, error)
+	GetPack(ctx context.Context, contractAddress string, chainId uint64) (*Pack, error)
+	DeletePack(ctx context.Context, contractAddress string, chainId uint64) (bool, error)
+	UpdatePackContent(ctx context.Context, pack *Pack) (string, error)
 }
 
 //
@@ -1367,6 +1722,11 @@ type APIClient interface {
 	GetSardineNFTCheckoutToken(ctx context.Context, params *SardineNFTCheckoutParams) (*SardineNFTCheckout, error)
 	// Deprecated. Use SardineGetNFTCheckoutOrderStatus() instead.
 	GetSardineNFTCheckoutOrderStatus(ctx context.Context, orderId string) (*SardineOrder, error)
+	TransakGetCountries(ctx context.Context) ([]*TransakCountry, error)
+	TransakGetCryptoCurrencies(ctx context.Context) ([]*TransakCryptoCurrency, error)
+	TransakGetFiatCurrencies(ctx context.Context) ([]*TransakFiatCurrency, error)
+	TransakGetPrice(ctx context.Context, params *TransakGetPriceParams) (*TransakPrice, error)
+	TransakGetSupportedNFTCheckoutChains(ctx context.Context) ([]*TransakChain, error)
 	//
 	// Price Feed
 	//
@@ -1414,9 +1774,17 @@ type APIClient interface {
 	//
 	// Currency abstraction
 	//
+	GetSwapPermit2Price(ctx context.Context, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) (*SwapPermit2Price, error)
+	GetSwapPermit2Prices(ctx context.Context, userAddress string, buyCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) ([]*SwapPermit2Price, error)
+	GetSwapPermit2Quote(ctx context.Context, userAddress string, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, includeApprove bool, slippagePercentage *uint64) (*SwapPermit2Quote, error)
 	GetSwapPrice(ctx context.Context, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) (*SwapPrice, error)
 	GetSwapPrices(ctx context.Context, userAddress string, buyCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) ([]*SwapPrice, error)
 	GetSwapQuote(ctx context.Context, userAddress string, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, includeApprove bool, slippagePercentage *uint64) (*SwapQuote, error)
+	GetSwapQuoteV2(ctx context.Context, userAddress string, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, includeApprove bool, slippagePercentage *uint64) (*SwapQuote, error)
+	//
+	// Chain abstraction
+	//
+	IntentQuery(ctx context.Context, wallet string, preconditions []*IntentPrecondition) ([]*IntentSolution, error)
 	//
 	// Inventory, payments and management
 	//
@@ -1428,6 +1796,13 @@ type APIClient interface {
 	DeleteOffchainInventory(ctx context.Context, inventoryId uint64) (bool, error)
 	RequestOffchainPayment(ctx context.Context, inventoryId uint64, recipient string, chainId *uint64, tokenAddress *string) (*PaymentResponse, error)
 	ListOffchainPayments(ctx context.Context, inventoryId uint64, page *Page) (*Page, []*OffchainPayment, error)
+	//
+	// Packs
+	//
+	SavePack(ctx context.Context, pack *Pack) (string, error)
+	GetPack(ctx context.Context, contractAddress string, chainId uint64) (*Pack, error)
+	DeletePack(ctx context.Context, contractAddress string, chainId uint64) (bool, error)
+	UpdatePackContent(ctx context.Context, pack *Pack) (string, error)
 }
 
 //
@@ -1438,12 +1813,12 @@ const APIPathPrefix = "/rpc/API/"
 
 type aPIClient struct {
 	client HTTPClient
-	urls   [69]string
+	urls   [83]string
 }
 
 func NewAPIClient(addr string, client HTTPClient) APIClient {
 	prefix := urlBase(addr) + APIPathPrefix
-	urls := [69]string{
+	urls := [83]string{
 		prefix + "Ping",
 		prefix + "Version",
 		prefix + "RuntimeStatus",
@@ -1485,6 +1860,11 @@ func NewAPIClient(addr string, client HTTPClient) APIClient {
 		prefix + "GetSardineClientToken",
 		prefix + "GetSardineNFTCheckoutToken",
 		prefix + "GetSardineNFTCheckoutOrderStatus",
+		prefix + "TransakGetCountries",
+		prefix + "TransakGetCryptoCurrencies",
+		prefix + "TransakGetFiatCurrencies",
+		prefix + "TransakGetPrice",
+		prefix + "TransakGetSupportedNFTCheckoutChains",
 		prefix + "GetCoinPrices",
 		prefix + "GetCollectiblePrices",
 		prefix + "GetExchangeRate",
@@ -1502,9 +1882,14 @@ func NewAPIClient(addr string, client HTTPClient) APIClient {
 		prefix + "GenerateWaaSVerificationURL",
 		prefix + "ValidateWaaSVerificationNonce",
 		prefix + "ListAdoptedWallets",
+		prefix + "GetSwapPermit2Price",
+		prefix + "GetSwapPermit2Prices",
+		prefix + "GetSwapPermit2Quote",
 		prefix + "GetSwapPrice",
 		prefix + "GetSwapPrices",
 		prefix + "GetSwapQuote",
+		prefix + "GetSwapQuoteV2",
+		prefix + "IntentQuery",
 		prefix + "ListCurrencyGroups",
 		prefix + "AddOffchainInventory",
 		prefix + "GetOffchainInventory",
@@ -1513,6 +1898,10 @@ func NewAPIClient(addr string, client HTTPClient) APIClient {
 		prefix + "DeleteOffchainInventory",
 		prefix + "RequestOffchainPayment",
 		prefix + "ListOffchainPayments",
+		prefix + "SavePack",
+		prefix + "GetPack",
+		prefix + "DeletePack",
+		prefix + "UpdatePackContent",
 	}
 	return &aPIClient{
 		client: client,
@@ -2302,6 +2691,89 @@ func (c *aPIClient) GetSardineNFTCheckoutOrderStatus(ctx context.Context, orderI
 	return out.Ret0, err
 }
 
+func (c *aPIClient) TransakGetCountries(ctx context.Context) ([]*TransakCountry, error) {
+	out := struct {
+		Ret0 []*TransakCountry `json:"regions"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[41], nil, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) TransakGetCryptoCurrencies(ctx context.Context) ([]*TransakCryptoCurrency, error) {
+	out := struct {
+		Ret0 []*TransakCryptoCurrency `json:"currencies"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[42], nil, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) TransakGetFiatCurrencies(ctx context.Context) ([]*TransakFiatCurrency, error) {
+	out := struct {
+		Ret0 []*TransakFiatCurrency `json:"currencies"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[43], nil, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) TransakGetPrice(ctx context.Context, params *TransakGetPriceParams) (*TransakPrice, error) {
+	in := struct {
+		Arg0 *TransakGetPriceParams `json:"params"`
+	}{params}
+	out := struct {
+		Ret0 *TransakPrice `json:"price"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[44], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) TransakGetSupportedNFTCheckoutChains(ctx context.Context) ([]*TransakChain, error) {
+	out := struct {
+		Ret0 []*TransakChain `json:"chains"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[45], nil, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
 func (c *aPIClient) GetCoinPrices(ctx context.Context, tokens []*Token) ([]*TokenPrice, error) {
 	in := struct {
 		Arg0 []*Token `json:"tokens"`
@@ -2310,7 +2782,7 @@ func (c *aPIClient) GetCoinPrices(ctx context.Context, tokens []*Token) ([]*Toke
 		Ret0 []*TokenPrice `json:"tokenPrices"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[41], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[46], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2329,7 +2801,7 @@ func (c *aPIClient) GetCollectiblePrices(ctx context.Context, tokens []*Token) (
 		Ret0 []*TokenPrice `json:"tokenPrices"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[42], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[47], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2348,7 +2820,7 @@ func (c *aPIClient) GetExchangeRate(ctx context.Context, toCurrency string) (*Ex
 		Ret0 *ExchangeRate `json:"exchangeRate"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[43], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[48], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2368,7 +2840,7 @@ func (c *aPIClient) MemoryStore(ctx context.Context, key string, value string) (
 		Ret0 bool `json:"ok"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[44], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[49], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2387,7 +2859,7 @@ func (c *aPIClient) MemoryLoad(ctx context.Context, key string) (string, error) 
 		Ret0 string `json:"value"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[45], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[50], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2403,7 +2875,7 @@ func (c *aPIClient) GetInviteInfo(ctx context.Context) (*InviteInfo, error) {
 		Ret0 *InviteInfo `json:"inviteInfo"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[46], nil, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[51], nil, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2422,7 +2894,7 @@ func (c *aPIClient) IsValidAccessCode(ctx context.Context, accessCode string) (b
 		Ret0 bool `json:"status"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[47], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[52], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2442,7 +2914,7 @@ func (c *aPIClient) InternalClaimAccessCode(ctx context.Context, address string,
 		Ret0 bool `json:"status"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[48], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[53], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2462,7 +2934,7 @@ func (c *aPIClient) BlockNumberAtTime(ctx context.Context, chainId uint64, times
 		Ret0 []uint64 `json:"blocks"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[49], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[54], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2484,7 +2956,7 @@ func (c *aPIClient) PaperSessionSecret(ctx context.Context, chainName string, co
 		Ret0 string `json:"secret"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[50], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[55], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2506,7 +2978,7 @@ func (c *aPIClient) PaperSessionSecret2(ctx context.Context, chainName string, c
 		Ret0 string `json:"secret"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[51], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[56], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2532,7 +3004,7 @@ func (c *aPIClient) LinkWallet(ctx context.Context, parentWalletAddress string, 
 		Ret0 bool `json:"status"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[52], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[57], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2554,7 +3026,7 @@ func (c *aPIClient) GetLinkedWallets(ctx context.Context, parentWalletAddress st
 		Ret0 []*LinkedWallet `json:"linkedWallets"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[53], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[58], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2577,7 +3049,7 @@ func (c *aPIClient) RemoveLinkedWallet(ctx context.Context, parentWalletAddress 
 		Ret0 bool `json:"status"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[54], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[59], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2597,7 +3069,7 @@ func (c *aPIClient) GenerateWaaSVerificationURL(ctx context.Context, walletAddre
 		Ret1 string `json:"verificationURL"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[55], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[60], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2619,7 +3091,7 @@ func (c *aPIClient) ValidateWaaSVerificationNonce(ctx context.Context, nonce str
 		Ret0 string `json:"walletAddress"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[56], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[61], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2639,7 +3111,7 @@ func (c *aPIClient) ListAdoptedWallets(ctx context.Context, page *Page) (*Page, 
 		Ret1 []*AdoptedChildWallet `json:"wallets"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[57], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[62], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2648,6 +3120,77 @@ func (c *aPIClient) ListAdoptedWallets(ctx context.Context, page *Page) (*Page, 
 	}
 
 	return out.Ret0, out.Ret1, err
+}
+
+func (c *aPIClient) GetSwapPermit2Price(ctx context.Context, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) (*SwapPermit2Price, error) {
+	in := struct {
+		Arg0 string  `json:"buyCurrencyAddress"`
+		Arg1 string  `json:"sellCurrencyAddress"`
+		Arg2 string  `json:"buyAmount"`
+		Arg3 uint64  `json:"chainId"`
+		Arg4 *uint64 `json:"slippagePercentage"`
+	}{buyCurrencyAddress, sellCurrencyAddress, buyAmount, chainId, slippagePercentage}
+	out := struct {
+		Ret0 *SwapPermit2Price `json:"swapPermit2Price"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[63], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) GetSwapPermit2Prices(ctx context.Context, userAddress string, buyCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) ([]*SwapPermit2Price, error) {
+	in := struct {
+		Arg0 string  `json:"userAddress"`
+		Arg1 string  `json:"buyCurrencyAddress"`
+		Arg2 string  `json:"buyAmount"`
+		Arg3 uint64  `json:"chainId"`
+		Arg4 *uint64 `json:"slippagePercentage"`
+	}{userAddress, buyCurrencyAddress, buyAmount, chainId, slippagePercentage}
+	out := struct {
+		Ret0 []*SwapPermit2Price `json:"swapPermit2Prices"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[64], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) GetSwapPermit2Quote(ctx context.Context, userAddress string, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, includeApprove bool, slippagePercentage *uint64) (*SwapPermit2Quote, error) {
+	in := struct {
+		Arg0 string  `json:"userAddress"`
+		Arg1 string  `json:"buyCurrencyAddress"`
+		Arg2 string  `json:"sellCurrencyAddress"`
+		Arg3 string  `json:"buyAmount"`
+		Arg4 uint64  `json:"chainId"`
+		Arg5 bool    `json:"includeApprove"`
+		Arg6 *uint64 `json:"slippagePercentage"`
+	}{userAddress, buyCurrencyAddress, sellCurrencyAddress, buyAmount, chainId, includeApprove, slippagePercentage}
+	out := struct {
+		Ret0 *SwapPermit2Quote `json:"swapPermit2Quote"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[65], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
 }
 
 func (c *aPIClient) GetSwapPrice(ctx context.Context, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, slippagePercentage *uint64) (*SwapPrice, error) {
@@ -2662,7 +3205,7 @@ func (c *aPIClient) GetSwapPrice(ctx context.Context, buyCurrencyAddress string,
 		Ret0 *SwapPrice `json:"swapPrice"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[58], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[66], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2685,7 +3228,7 @@ func (c *aPIClient) GetSwapPrices(ctx context.Context, userAddress string, buyCu
 		Ret0 []*SwapPrice `json:"swapPrices"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[59], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[67], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2710,7 +3253,52 @@ func (c *aPIClient) GetSwapQuote(ctx context.Context, userAddress string, buyCur
 		Ret0 *SwapQuote `json:"swapQuote"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[60], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[68], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) GetSwapQuoteV2(ctx context.Context, userAddress string, buyCurrencyAddress string, sellCurrencyAddress string, buyAmount string, chainId uint64, includeApprove bool, slippagePercentage *uint64) (*SwapQuote, error) {
+	in := struct {
+		Arg0 string  `json:"userAddress"`
+		Arg1 string  `json:"buyCurrencyAddress"`
+		Arg2 string  `json:"sellCurrencyAddress"`
+		Arg3 string  `json:"buyAmount"`
+		Arg4 uint64  `json:"chainId"`
+		Arg5 bool    `json:"includeApprove"`
+		Arg6 *uint64 `json:"slippagePercentage"`
+	}{userAddress, buyCurrencyAddress, sellCurrencyAddress, buyAmount, chainId, includeApprove, slippagePercentage}
+	out := struct {
+		Ret0 *SwapQuote `json:"swapQuote"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[69], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) IntentQuery(ctx context.Context, wallet string, preconditions []*IntentPrecondition) ([]*IntentSolution, error) {
+	in := struct {
+		Arg0 string                `json:"wallet"`
+		Arg1 []*IntentPrecondition `json:"preconditions"`
+	}{wallet, preconditions}
+	out := struct {
+		Ret0 []*IntentSolution `json:"solutions"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[70], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2726,7 +3314,7 @@ func (c *aPIClient) ListCurrencyGroups(ctx context.Context) ([]*CurrencyGroup, e
 		Ret0 []*CurrencyGroup `json:"currencyGroups"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[61], nil, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[71], nil, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2745,7 +3333,7 @@ func (c *aPIClient) AddOffchainInventory(ctx context.Context, inventory *Offchai
 		Ret0 uint64 `json:"inventoryId"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[62], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[72], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2764,7 +3352,7 @@ func (c *aPIClient) GetOffchainInventory(ctx context.Context, inventoryId uint64
 		Ret0 *OffchainInventory `json:"inventory"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[63], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[73], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2783,7 +3371,7 @@ func (c *aPIClient) ListOffchainInventories(ctx context.Context, projectId uint6
 		Ret0 []*OffchainInventory `json:"inventory"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[64], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[74], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2799,7 +3387,7 @@ func (c *aPIClient) UpdateOffchainInventory(ctx context.Context, inventory *Offc
 		Arg0 *OffchainInventory `json:"inventory"`
 	}{inventory}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[65], in, nil)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[75], in, nil)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2818,7 +3406,7 @@ func (c *aPIClient) DeleteOffchainInventory(ctx context.Context, inventoryId uin
 		Ret0 bool `json:"ok"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[66], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[76], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2840,7 +3428,7 @@ func (c *aPIClient) RequestOffchainPayment(ctx context.Context, inventoryId uint
 		Ret0 *PaymentResponse `json:"payment"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[67], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[77], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2861,7 +3449,7 @@ func (c *aPIClient) ListOffchainPayments(ctx context.Context, inventoryId uint64
 		Ret1 []*OffchainPayment `json:"payments"`
 	}{}
 
-	resp, err := doHTTPRequest(ctx, c.client, c.urls[68], in, &out)
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[78], in, &out)
 	if resp != nil {
 		cerr := resp.Body.Close()
 		if err == nil && cerr != nil {
@@ -2870,6 +3458,84 @@ func (c *aPIClient) ListOffchainPayments(ctx context.Context, inventoryId uint64
 	}
 
 	return out.Ret0, out.Ret1, err
+}
+
+func (c *aPIClient) SavePack(ctx context.Context, pack *Pack) (string, error) {
+	in := struct {
+		Arg0 *Pack `json:"pack"`
+	}{pack}
+	out := struct {
+		Ret0 string `json:"merkleRoot"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[79], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) GetPack(ctx context.Context, contractAddress string, chainId uint64) (*Pack, error) {
+	in := struct {
+		Arg0 string `json:"contractAddress"`
+		Arg1 uint64 `json:"chainId"`
+	}{contractAddress, chainId}
+	out := struct {
+		Ret0 *Pack `json:"pack"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[80], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) DeletePack(ctx context.Context, contractAddress string, chainId uint64) (bool, error) {
+	in := struct {
+		Arg0 string `json:"contractAddress"`
+		Arg1 uint64 `json:"chainId"`
+	}{contractAddress, chainId}
+	out := struct {
+		Ret0 bool `json:"status"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[81], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
+}
+
+func (c *aPIClient) UpdatePackContent(ctx context.Context, pack *Pack) (string, error) {
+	in := struct {
+		Arg0 *Pack `json:"pack"`
+	}{pack}
+	out := struct {
+		Ret0 string `json:"merkleRoot"`
+	}{}
+
+	resp, err := doHTTPRequest(ctx, c.client, c.urls[82], in, &out)
+	if resp != nil {
+		cerr := resp.Body.Close()
+		if err == nil && cerr != nil {
+			err = ErrWebrpcRequestFailed.WithCausef("failed to close response body: %w", cerr)
+		}
+	}
+
+	return out.Ret0, err
 }
 
 // HTTPClient is the interface used by generated clients to send HTTP requests.
