@@ -1,8 +1,13 @@
 package main
 
 import (
-	"encoding/json"
+	"errors"
 	"fmt"
+	"math/big"
+	"strconv"
+	"strings"
+
+	"encoding/hex"
 
 	"github.com/0xsequence/ethkit/go-ethereum/common"
 	v3 "github.com/0xsequence/go-sequence/core/v3"
@@ -10,39 +15,83 @@ import (
 )
 
 func handleEmptyTopology(p *EmptyTopologyParams) (string, error) {
-	signer := common.HexToAddress(p.GlobalSigner)
+	signer := common.HexToAddress(p.IdentitySigner)
 	topology := v3.EmptySessionsTopology(signer)
 	return v3.SessionsTopologyToJSON(topology)
 }
 
-func handleEncodeConfiguration(p *EncodeConfigurationParams) (string, error) {
-	topology, err := v3.SessionsTopologyFromJSON(p.SessionConfiguration)
+func handleEncodeTopology(p *EncodeTopologyParams) (string, error) {
+	encoded, err := v3.EncodeSessionsTopology(p.SessionTopology)
 	if err != nil {
-		return "", fmt.Errorf("failed to decode session configuration: %w", err)
+		return "", fmt.Errorf("failed to encode session topology: %w", err)
 	}
-
-	configTree := v3.SessionsTopologyToConfigurationTree(topology)
-	jsonBytes, err := json.Marshal(configTree)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal configuration tree: %w", err)
-	}
-
-	return string(jsonBytes), nil
+	return "0x" + common.Bytes2Hex(encoded), nil
 }
 
 func handleEncodeSessionCallSignatures(p *EncodeSessionCallSignaturesParams) (string, error) {
-	topology, err := v3.SessionsTopologyFromJSON(p.SessionConfiguration)
-	if err != nil {
-		return "", fmt.Errorf("failed to decode session configuration: %w", err)
-	}
-
 	callSigs := make([]v3.SessionCallSignature, len(p.CallSignatures))
 	for i, sig := range p.CallSignatures {
-		callSig, err := v3.SessionCallSignatureFromJSON(sig)
+		sessionSig, err := rsyFromRsvStr(sig.SessionSignature)
 		if err != nil {
-			return "", fmt.Errorf("failed to decode call signature %d: %w", i, err)
+			return "", fmt.Errorf("failed to decode session signature: %w", err)
 		}
-		callSigs[i] = *callSig
+		if sig.Attestation.ApprovedSigner != "" && sig.IdentitySignature != "" {
+			// Implicit call signature
+			idSig, err := rsyFromRsvStr(sig.IdentitySignature)
+			if err != nil {
+				return "", fmt.Errorf("failed to decode identity signature: %w", err)
+			}
+
+			// Convert hex-encoded fields to bytes
+			identityType, err := hex.DecodeString(sig.Attestation.IdentityType[2:]) // Remove "0x" prefix
+			if err != nil {
+				return "", fmt.Errorf("failed to decode identity type: %w", err)
+			}
+
+			issuerHash, err := hex.DecodeString(sig.Attestation.IssuerHash[2:])
+			if err != nil {
+				return "", fmt.Errorf("failed to decode issuer hash: %w", err)
+			}
+
+			audienceHash, err := hex.DecodeString(sig.Attestation.AudienceHash[2:])
+			if err != nil {
+				return "", fmt.Errorf("failed to decode audience hash: %w", err)
+			}
+
+			applicationData, err := hex.DecodeString(sig.Attestation.ApplicationData[2:])
+			if err != nil {
+				return "", fmt.Errorf("failed to decode application data: %w", err)
+			}
+
+			attestation := v3.Attestation{
+				ApprovedSigner:  common.HexToAddress(sig.Attestation.ApprovedSigner),
+				IdentityType:    identityType,
+				IssuerHash:      issuerHash,
+				AudienceHash:    audienceHash,
+				ApplicationData: applicationData,
+				AuthData: v3.AuthData{
+					RedirectUrl: sig.Attestation.AuthData.RedirectUrl,
+				},
+			}
+
+			callSig := v3.ImplicitSessionCallSignature{
+				Attestation:       attestation,
+				IdentitySignature: idSig,
+				SessionSignature:  sessionSig,
+			}
+			callSigs[i] = callSig
+		} else {
+			// Explicit call signature
+			permissionIndex, ok := new(big.Int).SetString(sig.PermissionIndex, 10)
+			if !ok {
+				return "", fmt.Errorf("failed to decode permission index: %w", err)
+			}
+			callSig := v3.ExplicitSessionCallSignature{
+				PermissionIndex:  permissionIndex,
+				SessionSignature: sessionSig,
+			}
+			callSigs[i] = callSig
+		}
 	}
 
 	explicitAddrs := make([]common.Address, len(p.ExplicitSigners))
@@ -55,23 +104,43 @@ func handleEncodeSessionCallSignatures(p *EncodeSessionCallSignaturesParams) (st
 		implicitAddrs[i] = common.HexToAddress(signer)
 	}
 
-	encoded, err := v3.EncodeSessionCallSignatures(callSigs, topology, explicitAddrs, implicitAddrs)
+	encoded, err := v3.EncodeSessionCallSignatures(callSigs, p.SessionTopology, explicitAddrs, implicitAddrs)
 	if err != nil {
 		return "", fmt.Errorf("failed to encode session call signatures: %w", err)
 	}
-
-	return common.Bytes2Hex(encoded), nil
+	return "0x" + common.Bytes2Hex(encoded), nil
 }
 
 func handleImageHash(p *ImageHashParams) (string, error) {
-	topology, err := v3.SessionsTopologyFromJSON(p.SessionConfiguration)
+	configTree, err := v3.SessionsTopologyToGenericTree(p.SessionTopology)
 	if err != nil {
-		return "", fmt.Errorf("failed to decode session configuration: %w", err)
+		return "", fmt.Errorf("failed to convert session topology to generic tree: %w", err)
 	}
+	hash, err := v3.HashTree(configTree)
+	if err != nil {
+		return "", fmt.Errorf("failed to hash generic tree: %w", err)
+	}
+	return "0x" + common.Bytes2Hex(hash), nil
+}
 
-	configTree := v3.SessionsTopologyToConfigurationTree(topology)
-	hash := v3.HashConfigurationTree(configTree)
-	return common.Bytes2Hex(hash[:]), nil
+func rsyFromRsvStr(sigStr string) (v3.RSY, error) {
+	parts := strings.Split(sigStr, ":")
+	if len(parts) != 3 {
+		return v3.RSY{}, errors.New("signature must be in r:s:v format")
+	}
+	r, err := hex.DecodeString(parts[0][2:])
+	if err != nil {
+		return v3.RSY{}, fmt.Errorf("invalid r value: %w", err)
+	}
+	s, err := hex.DecodeString(parts[1][2:])
+	if err != nil {
+		return v3.RSY{}, fmt.Errorf("invalid s value: %w", err)
+	}
+	v, err := strconv.Atoi(string(parts[2]))
+	if err != nil {
+		return v3.RSY{}, fmt.Errorf("invalid v value: %w", err)
+	}
+	return v3.RSY{R: r, S: s, YParity: uint8(v - 27)}, nil
 }
 
 // newSessionCmd defines the command-line interface.
@@ -82,14 +151,14 @@ func newSessionCmd() *cobra.Command {
 	}
 
 	emptyCmd := &cobra.Command{
-		Use:   "empty [global-signer]",
-		Short: "Create an empty session topology with the given global signer",
+		Use:   "empty [identity-signer]",
+		Short: "Create an empty session topology with the given identity signer",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			globalSigner, err := fromPosOrStdin(args, 0)
+			identitySigner, err := fromPosOrStdin(args, 0)
 			if err != nil {
 				return err
 			}
-			result, err := handleEmptyTopology(&EmptyTopologyParams{GlobalSigner: globalSigner})
+			result, err := handleEmptyTopology(&EmptyTopologyParams{IdentitySigner: identitySigner})
 			if err != nil {
 				return err
 			}
@@ -98,15 +167,19 @@ func newSessionCmd() *cobra.Command {
 		},
 	}
 
-	encodeConfigurationCmd := &cobra.Command{
-		Use:   "encode-configuration [session-configuration]",
-		Short: "Encode a session configuration",
+	encodeTopologyCmd := &cobra.Command{
+		Use:   "encode-topology [session-topology]",
+		Short: "Encode a session topology",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			config, err := fromPosOrStdin(args, 0)
 			if err != nil {
 				return err
 			}
-			result, err := handleEncodeConfiguration(&EncodeConfigurationParams{SessionConfiguration: config})
+			topology, err := v3.SessionsTopologyFromJSON(config)
+			if err != nil {
+				return err
+			}
+			result, err := handleEncodeTopology(&EncodeTopologyParams{SessionTopology: topology})
 			if err != nil {
 				return err
 			}
@@ -116,23 +189,33 @@ func newSessionCmd() *cobra.Command {
 	}
 
 	encodeCallsCmd := &cobra.Command{
-		Use:   "encode-calls [session-configuration] [call-signatures...]",
-		Short: "Encode a call signature for an implicit session",
+		Use:   "encode-calls [session-topology] [call-signatures] [explicit-signers] [implicit-signers]",
+		Short: "Encode call signatures for sessions",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) < 2 {
-				return fmt.Errorf("requires session configuration and at least one call signature")
+			if len(args) < 4 {
+				return fmt.Errorf("requires session topology, call signatures, explicit signers, and implicit signers")
 			}
 
 			config := args[0]
-			callSigs := args[1:]
+			callSigs := make([]CallSignaturesParams, len(args)-1)
+			for i, sig := range args[1:] {
+				callSigs[i] = CallSignaturesParams{
+					SessionSignature: sig,
+				}
+			}
 			explicitSigners, _ := cmd.Flags().GetStringSlice("explicit-signers")
 			implicitSigners, _ := cmd.Flags().GetStringSlice("implicit-signers")
 
+			topology, err := v3.SessionsTopologyFromJSON(config)
+			if err != nil {
+				return err
+			}
+
 			result, err := handleEncodeSessionCallSignatures(&EncodeSessionCallSignaturesParams{
-				SessionConfiguration: config,
-				CallSignatures:       callSigs,
-				ExplicitSigners:      explicitSigners,
-				ImplicitSigners:      implicitSigners,
+				SessionTopology: topology,
+				CallSignatures:  callSigs,
+				ExplicitSigners: explicitSigners,
+				ImplicitSigners: implicitSigners,
 			})
 			if err != nil {
 				return err
@@ -145,14 +228,18 @@ func newSessionCmd() *cobra.Command {
 	encodeCallsCmd.Flags().StringSlice("implicit-signers", []string{}, "Implicit signers")
 
 	imageHashCmd := &cobra.Command{
-		Use:   "image-hash [session-configuration]",
-		Short: "Hash a session configuration",
+		Use:   "image-hash [session-topology]",
+		Short: "Hash a session topology",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			config, err := fromPosOrStdin(args, 0)
 			if err != nil {
 				return err
 			}
-			result, err := handleImageHash(&ImageHashParams{SessionConfiguration: config})
+			topology, err := v3.SessionsTopologyFromJSON(config)
+			if err != nil {
+				return err
+			}
+			result, err := handleImageHash(&ImageHashParams{SessionTopology: topology})
 			if err != nil {
 				return err
 			}
@@ -161,7 +248,7 @@ func newSessionCmd() *cobra.Command {
 		},
 	}
 
-	cmd.AddCommand(emptyCmd, encodeConfigurationCmd, encodeCallsCmd, imageHashCmd)
+	cmd.AddCommand(emptyCmd, encodeTopologyCmd, encodeCallsCmd, imageHashCmd)
 	cmd.AddCommand(newSessionExplicitCmd(), newSessionImplicitCmd())
 	return cmd
 }
