@@ -31,8 +31,8 @@ var (
 
 // WalletGuestMetaData contains all meta data concerning the WalletGuest contract.
 var WalletGuestMetaData = &bind.MetaData{
-	ABI: "[{\"inputs\":[{\"internalType\":\"uint256\",\"name\":\"index\",\"type\":\"uint256\"}],\"name\":\"DelegateCallNotAllowed\",\"type\":\"error\"},{\"inputs\":[{\"components\":[{\"internalType\":\"uint8\",\"name\":\"kind\",\"type\":\"uint8\"},{\"internalType\":\"bool\",\"name\":\"noChainId\",\"type\":\"bool\"},{\"components\":[{\"internalType\":\"address\",\"name\":\"to\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"data\",\"type\":\"bytes\"},{\"internalType\":\"uint256\",\"name\":\"gasLimit\",\"type\":\"uint256\"},{\"internalType\":\"bool\",\"name\":\"delegateCall\",\"type\":\"bool\"},{\"internalType\":\"bool\",\"name\":\"onlyFallback\",\"type\":\"bool\"},{\"internalType\":\"uint256\",\"name\":\"behaviorOnError\",\"type\":\"uint256\"}],\"internalType\":\"structPayload.Call[]\",\"name\":\"calls\",\"type\":\"tuple[]\"},{\"internalType\":\"uint256\",\"name\":\"space\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"nonce\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"message\",\"type\":\"bytes\"},{\"internalType\":\"bytes32\",\"name\":\"imageHash\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"digest\",\"type\":\"bytes32\"},{\"internalType\":\"address[]\",\"name\":\"parentWallets\",\"type\":\"address[]\"}],\"internalType\":\"structPayload.Decoded\",\"name\":\"_payload\",\"type\":\"tuple\"},{\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"_gasLeft\",\"type\":\"uint256\"}],\"name\":\"NotEnoughGas\",\"type\":\"error\"},{\"inputs\":[{\"components\":[{\"internalType\":\"uint8\",\"name\":\"kind\",\"type\":\"uint8\"},{\"internalType\":\"bool\",\"name\":\"noChainId\",\"type\":\"bool\"},{\"components\":[{\"internalType\":\"address\",\"name\":\"to\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"data\",\"type\":\"bytes\"},{\"internalType\":\"uint256\",\"name\":\"gasLimit\",\"type\":\"uint256\"},{\"internalType\":\"bool\",\"name\":\"delegateCall\",\"type\":\"bool\"},{\"internalType\":\"bool\",\"name\":\"onlyFallback\",\"type\":\"bool\"},{\"internalType\":\"uint256\",\"name\":\"behaviorOnError\",\"type\":\"uint256\"}],\"internalType\":\"structPayload.Call[]\",\"name\":\"calls\",\"type\":\"tuple[]\"},{\"internalType\":\"uint256\",\"name\":\"space\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"nonce\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"message\",\"type\":\"bytes\"},{\"internalType\":\"bytes32\",\"name\":\"imageHash\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"digest\",\"type\":\"bytes32\"},{\"internalType\":\"address[]\",\"name\":\"parentWallets\",\"type\":\"address[]\"}],\"internalType\":\"structPayload.Decoded\",\"name\":\"_payload\",\"type\":\"tuple\"},{\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"_returnData\",\"type\":\"bytes\"}],\"name\":\"Reverted\",\"type\":\"error\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"_opHash\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"}],\"name\":\"CallAborted\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"_opHash\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"}],\"name\":\"CallFailed\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"_opHash\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"}],\"name\":\"CallSkipped\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"_opHash\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"}],\"name\":\"CallSuccess\",\"type\":\"event\"},{\"stateMutability\":\"nonpayable\",\"type\":\"fallback\"}]",
-	Bin: "0x6080604052348015600f57600080fd5b5061111b8061001f6000396000f3fe608060405234801561001057600080fd5b503660008036600081811061002757610027610d35565b909101357fff00000000000000000000000000000000000000000000000000000000000000167f1f00000000000000000000000000000000000000000000000000000000000000039050610082575060649050602435610089565b5060009050365b600061009583836100b0565b905060006100a2826104ce565b90506100ae8282610549565b005b604080516101208101825260008082526020820181905260609282018390528282018190526080820181905260a0820183905260c0820181905260e0820152610100810191909152823560f81c600180821681036101145760006060840152610125565b84810135606090811c908401526014015b6007600183901c1680156101785760016008820290811b7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0187840135610100929092039190911c166080850152908101905b60008360101660100361018d575060016101b6565b836020166020036101a95750600282019186013560f01c6101b6565b50600182019186013560f81c5b8067ffffffffffffffff8111156101cf576101cf610d64565b60405190808252806020026020018201604052801561025a57816020015b6102476040518060e00160405280600073ffffffffffffffffffffffffffffffffffffffff168152602001600081526020016060815260200160008152602001600015158152602001600015158152602001600081525090565b8152602001906001900390816101ed5790505b50604086015260005b818110156104c35760018085019489013560f81c9080821690036102c257308760400151838151811061029857610298610d35565b602090810291909101015173ffffffffffffffffffffffffffffffffffffffff909116905261030c565b8489013560601c60148601886040015184815181106102e3576102e3610d35565b602090810291909101015173ffffffffffffffffffffffffffffffffffffffff90921690915294505b6002808216900361034a5784890135602086018860400151848151811061033557610335610d35565b60200260200101516020018197508281525050505b600480821690036103e257600385019489013560e81c89868a61036d8483610d93565b9261037a93929190610dcd565b8080601f01602080910402602001604051908101604052809392919081815260200183838082843760009201919091525050505060408901518051859081106103c5576103c5610d35565b6020908102919091010151604001526103de8187610d93565b9550505b600880821690036104205784890135602086018860400151848151811061040b5761040b610d35565b60200260200101516060018197508281525050505b8060101660ff166010148760400151838151811061044057610440610d35565b602002602001015160800190151590811515815250508060201660ff166020148760400151838151811061047657610476610d35565b602090810291909101015190151560a090910152604087015180516003600684901c169190849081106104ab576104ab610d35565b602090810291909101015160c0015250600101610263565b505050505092915050565b6000806104df83602001513061080a565b905060006104ec846108d7565b6040517f1901000000000000000000000000000000000000000000000000000000000000602082015260228101849052604281018290529091506062015b6040516020818303038152906040528051906020012092505050919050565b604082015151600090815b818110156108035760008560400151828151811061057457610574610d35565b602002602001015190508060a00151801561058d575083155b156105d5576040805186815260208101849052600095507f9ae934bf8a986157c889a24c3b3fa85e74b7e4ee4b1f8fc6e7362cb4c1d19d8b910160405180910390a1506107fb565b606081015180158015906105e85750805a105b1561062e5786835a6040517f2139527400000000000000000000000000000000000000000000000000000000815260040161062593929190611031565b60405180910390fd5b81608001511561066d576040517f230d1ccc00000000000000000000000000000000000000000000000000000000815260048101849052602401610625565b815160208301516000916106939184156106875784610689565b5a5b8660400151610b2d565b9050806107be5760c08301516106e8576040805188815260208101869052600197507f20832642214d5218c6428e71d8d2ddd9ad15a81ad2be8154d8c2e3ab08293fcb910160405180910390a15050506107fb565b60c08301517fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0161075257878461071d610b45565b6040517f7f6b0bb100000000000000000000000000000000000000000000000000000000815260040161062593929190611056565b60c08301517ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe016107be5760408051888152602081018690527f5b5cb72c79981de49f1f950d4d8d62397e2fc2b772e1b788a640025075ab47c3910160405180910390a1505050610803565b60408051888152602081018690527fec670aed5ee1e72eb3eb601271be4b3f312e71f17eebdf10c1a0ab5a3af30ffd910160405180910390a15050505b600101610554565b5050505050565b60007f8b73c3c69bb8fe3d512ecc4cf759cc79239f7b179b0ffacaa9a75d522b39400f7f4aa45ca7ad825ceb1bf35643f0a58c295239df563b1b565c2485f96477c563187f2a80e1ef1d7842f27f2e6be0972bb708b9a135c38860dbe73c27c3486c34f4de8561087a574661087d565b60005b6040805160208101959095528401929092526060830152608082015273ffffffffffffffffffffffffffffffffffffffff831660a082015260c0016040516020818303038152906040528051906020012090505b92915050565b6000806108e8836101000151610b64565b835190915060ff1661095b5760006109038460400151610be7565b606080860151608080880151604080517f11e1e4079a79a66e4ade50033cfe2678cdd5341d2dfe5ef9513edb1a0be147a260208201529081018690529384019290925282015260a0810184905290915060c00161052a565b825160ff167fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff016109eb5760a08301518051602091820120604080517fe19a3b94fc3c7ece3f890d98a99bc422615537a08dea0603fa8425867d87d46693810193909352820152606081018290526080015b60405160208183030381529060405280519060200120915050919050565b825160ff167ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe01610a5b5760c0830151604080517f11fdeb7e8373a1aa96bfac8d0ea91526b2c5d15e5cee20e0543e780258f3e8e4602082015290810191909152606081018290526080016109cd565b825160ff167ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffd01610acb5760e0830151604080517f402e923b91e918306019e73f589362164a6a059499a504699c66feabbb3e2624602082015290810191909152606081018290526080016109cd565b6040517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601060248201527f556e737570706f72746564206b696e64000000000000000000000000000000006044820152606401610625565b6000806000835160208501878988f195945050505050565b60603d604051915060208201818101604052818352816000823e505090565b6000606060005b8351811015610bd85781848281518110610b8757610b87610d35565b6020026020010151604051602001610ba092919061108b565b604080517fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe08184030181529190529150600101610b6b565b50805160209091012092915050565b6000606060005b8351811015610bd8576000610c1b858381518110610c0e57610c0e610d35565b6020026020010151610c69565b90508281604051602001610c309291906110c3565b604080517fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0818403018152919052925050600101610bee565b60007f0603985259a953da1f65a522f589c17bd1d0117ec1d3abb7c0788aef251ef43782600001518360200151846040015180519060200120856060015186608001518760a001518860c00151604051602001610d1898979695949392919097885273ffffffffffffffffffffffffffffffffffffffff969096166020880152604087019490945260608601929092526080850152151560a0840152151560c083015260e08201526101000190565b604051602081830303815290604052805190602001209050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052603260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b808201808211156108d1577f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b60008085851115610ddd57600080fd5b83861115610dea57600080fd5b5050820193919092039150565b60005b83811015610e12578181015183820152602001610dfa565b50506000910152565b60008151808452610e33816020860160208601610df7565b601f017fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0169290920160200192915050565b600082825180855260208501945060208160051b8301016020850160005b83811015610f35577fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0858403018852815173ffffffffffffffffffffffffffffffffffffffff815116845260208101516020850152604081015160e06040860152610ef160e0860182610e1b565b6060838101519087015260808084015115159087015260a08084015115159087015260c0928301519290950191909152506020978801979190910190600101610e83565b50909695505050505050565b600081518084526020840193506020830160005b82811015610f8957815173ffffffffffffffffffffffffffffffffffffffff16865260209586019590910190600101610f55565b5093949350505050565b805160ff16825260006020820151610faf602085018215159052565b5060408201516101206040850152610fcb610120850182610e65565b9050606083015160608501526080830151608085015260a083015184820360a0860152610ff88282610e1b565b91505060c083015160c085015260e083015160e08501526101008301518482036101008601526110288282610f41565b95945050505050565b6060815260006110446060830186610f93565b60208301949094525060400152919050565b6060815260006110696060830186610f93565b84602084015282810360408401526110818185610e1b565b9695505050505050565b60408152600061109e6040830185610e1b565b905073ffffffffffffffffffffffffffffffffffffffff831660208301529392505050565b600083516110d5818460208801610df7565b919091019182525060200191905056fea2646970667358221220b3ec6bd1eb613408c2f34b3fa302905c159ca9fe27b4a6eb199992f2b49b743a64736f6c634300081b0033",
+	ABI: "[{\"inputs\":[{\"internalType\":\"uint256\",\"name\":\"index\",\"type\":\"uint256\"}],\"name\":\"DelegateCallNotAllowed\",\"type\":\"error\"},{\"inputs\":[{\"components\":[{\"internalType\":\"uint8\",\"name\":\"kind\",\"type\":\"uint8\"},{\"internalType\":\"bool\",\"name\":\"noChainId\",\"type\":\"bool\"},{\"components\":[{\"internalType\":\"address\",\"name\":\"to\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"data\",\"type\":\"bytes\"},{\"internalType\":\"uint256\",\"name\":\"gasLimit\",\"type\":\"uint256\"},{\"internalType\":\"bool\",\"name\":\"delegateCall\",\"type\":\"bool\"},{\"internalType\":\"bool\",\"name\":\"onlyFallback\",\"type\":\"bool\"},{\"internalType\":\"uint256\",\"name\":\"behaviorOnError\",\"type\":\"uint256\"}],\"internalType\":\"structPayload.Call[]\",\"name\":\"calls\",\"type\":\"tuple[]\"},{\"internalType\":\"uint256\",\"name\":\"space\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"nonce\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"message\",\"type\":\"bytes\"},{\"internalType\":\"bytes32\",\"name\":\"imageHash\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"digest\",\"type\":\"bytes32\"},{\"internalType\":\"address[]\",\"name\":\"parentWallets\",\"type\":\"address[]\"}],\"internalType\":\"structPayload.Decoded\",\"name\":\"_payload\",\"type\":\"tuple\"},{\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"_gasLeft\",\"type\":\"uint256\"}],\"name\":\"NotEnoughGas\",\"type\":\"error\"},{\"inputs\":[{\"components\":[{\"internalType\":\"uint8\",\"name\":\"kind\",\"type\":\"uint8\"},{\"internalType\":\"bool\",\"name\":\"noChainId\",\"type\":\"bool\"},{\"components\":[{\"internalType\":\"address\",\"name\":\"to\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"data\",\"type\":\"bytes\"},{\"internalType\":\"uint256\",\"name\":\"gasLimit\",\"type\":\"uint256\"},{\"internalType\":\"bool\",\"name\":\"delegateCall\",\"type\":\"bool\"},{\"internalType\":\"bool\",\"name\":\"onlyFallback\",\"type\":\"bool\"},{\"internalType\":\"uint256\",\"name\":\"behaviorOnError\",\"type\":\"uint256\"}],\"internalType\":\"structPayload.Call[]\",\"name\":\"calls\",\"type\":\"tuple[]\"},{\"internalType\":\"uint256\",\"name\":\"space\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"nonce\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"message\",\"type\":\"bytes\"},{\"internalType\":\"bytes32\",\"name\":\"imageHash\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"digest\",\"type\":\"bytes32\"},{\"internalType\":\"address[]\",\"name\":\"parentWallets\",\"type\":\"address[]\"}],\"internalType\":\"structPayload.Decoded\",\"name\":\"_payload\",\"type\":\"tuple\"},{\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"},{\"internalType\":\"bytes\",\"name\":\"_returnData\",\"type\":\"bytes\"}],\"name\":\"Reverted\",\"type\":\"error\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"_opHash\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"},{\"indexed\":false,\"internalType\":\"bytes\",\"name\":\"_returnData\",\"type\":\"bytes\"}],\"name\":\"CallAborted\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"_opHash\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"},{\"indexed\":false,\"internalType\":\"bytes\",\"name\":\"_returnData\",\"type\":\"bytes\"}],\"name\":\"CallFailed\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"_opHash\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"}],\"name\":\"CallSkipped\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"_opHash\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"_index\",\"type\":\"uint256\"}],\"name\":\"CallSucceeded\",\"type\":\"event\"},{\"stateMutability\":\"nonpayable\",\"type\":\"fallback\"}]",
+	Bin: "0x6080604052348015600f57600080fd5b5061114c8061001f6000396000f3fe608060405234801561001057600080fd5b503660008036600081811061002757610027610d47565b909101357fff00000000000000000000000000000000000000000000000000000000000000167f1f00000000000000000000000000000000000000000000000000000000000000039050610082575060649050602435610089565b5060009050365b600061009583836100b0565b905060006100a2826104ce565b90506100ae8282610549565b005b604080516101208101825260008082526020820181905260609282018390528282018190526080820181905260a0820183905260c0820181905260e0820152610100810191909152823560f81c600180821681036101145760006060840152610125565b84810135606090811c908401526014015b6007600183901c1680156101785760016008820290811b7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0187840135610100929092039190911c166080850152908101905b60008360101660100361018d575060016101b6565b836020166020036101a95750600282019186013560f01c6101b6565b50600182019186013560f81c5b8067ffffffffffffffff8111156101cf576101cf610d76565b60405190808252806020026020018201604052801561025a57816020015b6102476040518060e00160405280600073ffffffffffffffffffffffffffffffffffffffff168152602001600081526020016060815260200160008152602001600015158152602001600015158152602001600081525090565b8152602001906001900390816101ed5790505b50604086015260005b818110156104c35760018085019489013560f81c9080821690036102c257308760400151838151811061029857610298610d47565b602090810291909101015173ffffffffffffffffffffffffffffffffffffffff909116905261030c565b8489013560601c60148601886040015184815181106102e3576102e3610d47565b602090810291909101015173ffffffffffffffffffffffffffffffffffffffff90921690915294505b6002808216900361034a5784890135602086018860400151848151811061033557610335610d47565b60200260200101516020018197508281525050505b600480821690036103e257600385019489013560e81c89868a61036d8483610da5565b9261037a93929190610ddf565b8080601f01602080910402602001604051908101604052809392919081815260200183838082843760009201919091525050505060408901518051859081106103c5576103c5610d47565b6020908102919091010151604001526103de8187610da5565b9550505b600880821690036104205784890135602086018860400151848151811061040b5761040b610d47565b60200260200101516060018197508281525050505b8060101660ff166010148760400151838151811061044057610440610d47565b602002602001015160800190151590811515815250508060201660ff166020148760400151838151811061047657610476610d47565b602090810291909101015190151560a090910152604087015180516003600684901c169190849081106104ab576104ab610d47565b602090810291909101015160c0015250600101610263565b505050505092915050565b6000806104df83602001513061081c565b905060006104ec846108e9565b6040517f1901000000000000000000000000000000000000000000000000000000000000602082015260228101849052604281018290529091506062015b6040516020818303038152906040528051906020012092505050919050565b604082015151600090815b818110156108155760008560400151828151811061057457610574610d47565b602002602001015190508060a00151801561058d575083155b156105d5576040805186815260208101849052600095507f9ae934bf8a986157c889a24c3b3fa85e74b7e4ee4b1f8fc6e7362cb4c1d19d8b910160405180910390a15061080d565b606081015180158015906105e85750805a105b1561062e5786835a6040517f2139527400000000000000000000000000000000000000000000000000000000815260040161062593929190611043565b60405180910390fd5b81608001511561066d576040517f230d1ccc00000000000000000000000000000000000000000000000000000000815260048101849052602401610625565b815160208301516000916106939184156106875784610689565b5a5b8660400151610b3f565b9050806107d05760c08301516106f157600195507f115f347c00e69f252cd6b63c4f81022a9564c6befe8aa719cb74640a4a306f0d87856106d2610b57565b6040516106e193929190611068565b60405180910390a150505061080d565b60c08301517fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0161075b578784610726610b57565b6040517f7f6b0bb100000000000000000000000000000000000000000000000000000000815260040161062593929190611087565b60c08301517ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe016107d0577fc2c704302430fe0dc8d95f272e2f4e54bbbc51a3327fd5d75ab41f9fc8fd129b87856107b1610b57565b6040516107c093929190611068565b60405180910390a1505050610815565b60408051888152602081018690527f5a589b1d8062f33451d29cae3dabd9b2e36c62aee644178c600977ca8dda661a910160405180910390a15050505b600101610554565b5050505050565b60007f8b73c3c69bb8fe3d512ecc4cf759cc79239f7b179b0ffacaa9a75d522b39400f7f4aa45ca7ad825ceb1bf35643f0a58c295239df563b1b565c2485f96477c563187f2a80e1ef1d7842f27f2e6be0972bb708b9a135c38860dbe73c27c3486c34f4de8561088c574661088f565b60005b6040805160208101959095528401929092526060830152608082015273ffffffffffffffffffffffffffffffffffffffff831660a082015260c0016040516020818303038152906040528051906020012090505b92915050565b6000806108fa836101000151610b76565b835190915060ff1661096d5760006109158460400151610bf9565b606080860151608080880151604080517f11e1e4079a79a66e4ade50033cfe2678cdd5341d2dfe5ef9513edb1a0be147a260208201529081018690529384019290925282015260a0810184905290915060c00161052a565b825160ff167fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff016109fd5760a08301518051602091820120604080517fe19a3b94fc3c7ece3f890d98a99bc422615537a08dea0603fa8425867d87d46693810193909352820152606081018290526080015b60405160208183030381529060405280519060200120915050919050565b825160ff167ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe01610a6d5760c0830151604080517f11fdeb7e8373a1aa96bfac8d0ea91526b2c5d15e5cee20e0543e780258f3e8e4602082015290810191909152606081018290526080016109df565b825160ff167ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffd01610add5760e0830151604080517fe19a3b94fc3c7ece3f890d98a99bc422615537a08dea0603fa8425867d87d466602082015290810191909152606081018290526080016109df565b6040517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601060248201527f556e737570706f72746564206b696e64000000000000000000000000000000006044820152606401610625565b6000806000835160208501878988f195945050505050565b60603d604051915060208201818101604052818352816000823e505090565b6000606060005b8351811015610bea5781848281518110610b9957610b99610d47565b6020026020010151604051602001610bb29291906110bc565b604080517fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe08184030181529190529150600101610b7d565b50805160209091012092915050565b6000606060005b8351811015610bea576000610c2d858381518110610c2057610c20610d47565b6020026020010151610c7b565b90508281604051602001610c429291906110f4565b604080517fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0818403018152919052925050600101610c00565b60007f0603985259a953da1f65a522f589c17bd1d0117ec1d3abb7c0788aef251ef43782600001518360200151846040015180519060200120856060015186608001518760a001518860c00151604051602001610d2a98979695949392919097885273ffffffffffffffffffffffffffffffffffffffff969096166020880152604087019490945260608601929092526080850152151560a0840152151560c083015260e08201526101000190565b604051602081830303815290604052805190602001209050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052603260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b808201808211156108e3577f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b60008085851115610def57600080fd5b83861115610dfc57600080fd5b5050820193919092039150565b60005b83811015610e24578181015183820152602001610e0c565b50506000910152565b60008151808452610e45816020860160208601610e09565b601f017fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0169290920160200192915050565b600082825180855260208501945060208160051b8301016020850160005b83811015610f47577fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0858403018852815173ffffffffffffffffffffffffffffffffffffffff815116845260208101516020850152604081015160e06040860152610f0360e0860182610e2d565b6060838101519087015260808084015115159087015260a08084015115159087015260c0928301519290950191909152506020978801979190910190600101610e95565b50909695505050505050565b600081518084526020840193506020830160005b82811015610f9b57815173ffffffffffffffffffffffffffffffffffffffff16865260209586019590910190600101610f67565b5093949350505050565b805160ff16825260006020820151610fc1602085018215159052565b5060408201516101206040850152610fdd610120850182610e77565b9050606083015160608501526080830151608085015260a083015184820360a086015261100a8282610e2d565b91505060c083015160c085015260e083015160e085015261010083015184820361010086015261103a8282610f53565b95945050505050565b6060815260006110566060830186610fa5565b60208301949094525060400152919050565b83815282602082015260606040820152600061103a6060830184610e2d565b60608152600061109a6060830186610fa5565b84602084015282810360408401526110b28185610e2d565b9695505050505050565b6040815260006110cf6040830185610e2d565b905073ffffffffffffffffffffffffffffffffffffffff831660208301529392505050565b60008351611106818460208801610e09565b919091019182525060200191905056fea264697066735822122069b92319a9ae77d4e1268d63cdb3377dbc078a37e6651ca3b8e11f003b4a88d864736f6c634300081b0033",
 }
 
 // WalletGuestABI is the input ABI used to generate the binding from.
@@ -292,14 +292,15 @@ func (it *WalletGuestCallAbortedIterator) Close() error {
 
 // WalletGuestCallAborted represents a CallAborted event raised by the WalletGuest contract.
 type WalletGuestCallAborted struct {
-	OpHash [32]byte
-	Index  *big.Int
-	Raw    types.Log // Blockchain specific contextual infos
+	OpHash     [32]byte
+	Index      *big.Int
+	ReturnData []byte
+	Raw        types.Log // Blockchain specific contextual infos
 }
 
-// FilterCallAborted is a free log retrieval operation binding the contract event 0x5b5cb72c79981de49f1f950d4d8d62397e2fc2b772e1b788a640025075ab47c3.
+// FilterCallAborted is a free log retrieval operation binding the contract event 0xc2c704302430fe0dc8d95f272e2f4e54bbbc51a3327fd5d75ab41f9fc8fd129b.
 //
-// Solidity: event CallAborted(bytes32 _opHash, uint256 _index)
+// Solidity: event CallAborted(bytes32 _opHash, uint256 _index, bytes _returnData)
 func (_WalletGuest *WalletGuestFilterer) FilterCallAborted(opts *bind.FilterOpts) (*WalletGuestCallAbortedIterator, error) {
 
 	logs, sub, err := _WalletGuest.contract.FilterLogs(opts, "CallAborted")
@@ -309,9 +310,9 @@ func (_WalletGuest *WalletGuestFilterer) FilterCallAborted(opts *bind.FilterOpts
 	return &WalletGuestCallAbortedIterator{contract: _WalletGuest.contract, event: "CallAborted", logs: logs, sub: sub}, nil
 }
 
-// WatchCallAborted is a free log subscription operation binding the contract event 0x5b5cb72c79981de49f1f950d4d8d62397e2fc2b772e1b788a640025075ab47c3.
+// WatchCallAborted is a free log subscription operation binding the contract event 0xc2c704302430fe0dc8d95f272e2f4e54bbbc51a3327fd5d75ab41f9fc8fd129b.
 //
-// Solidity: event CallAborted(bytes32 _opHash, uint256 _index)
+// Solidity: event CallAborted(bytes32 _opHash, uint256 _index, bytes _returnData)
 func (_WalletGuest *WalletGuestFilterer) WatchCallAborted(opts *bind.WatchOpts, sink chan<- *WalletGuestCallAborted) (event.Subscription, error) {
 
 	logs, sub, err := _WalletGuest.contract.WatchLogs(opts, "CallAborted")
@@ -346,9 +347,9 @@ func (_WalletGuest *WalletGuestFilterer) WatchCallAborted(opts *bind.WatchOpts, 
 	}), nil
 }
 
-// ParseCallAborted is a log parse operation binding the contract event 0x5b5cb72c79981de49f1f950d4d8d62397e2fc2b772e1b788a640025075ab47c3.
+// ParseCallAborted is a log parse operation binding the contract event 0xc2c704302430fe0dc8d95f272e2f4e54bbbc51a3327fd5d75ab41f9fc8fd129b.
 //
-// Solidity: event CallAborted(bytes32 _opHash, uint256 _index)
+// Solidity: event CallAborted(bytes32 _opHash, uint256 _index, bytes _returnData)
 func (_WalletGuest *WalletGuestFilterer) ParseCallAborted(log types.Log) (*WalletGuestCallAborted, error) {
 	event := new(WalletGuestCallAborted)
 	if err := _WalletGuest.contract.UnpackLog(event, "CallAborted", log); err != nil {
@@ -427,14 +428,15 @@ func (it *WalletGuestCallFailedIterator) Close() error {
 
 // WalletGuestCallFailed represents a CallFailed event raised by the WalletGuest contract.
 type WalletGuestCallFailed struct {
-	OpHash [32]byte
-	Index  *big.Int
-	Raw    types.Log // Blockchain specific contextual infos
+	OpHash     [32]byte
+	Index      *big.Int
+	ReturnData []byte
+	Raw        types.Log // Blockchain specific contextual infos
 }
 
-// FilterCallFailed is a free log retrieval operation binding the contract event 0x20832642214d5218c6428e71d8d2ddd9ad15a81ad2be8154d8c2e3ab08293fcb.
+// FilterCallFailed is a free log retrieval operation binding the contract event 0x115f347c00e69f252cd6b63c4f81022a9564c6befe8aa719cb74640a4a306f0d.
 //
-// Solidity: event CallFailed(bytes32 _opHash, uint256 _index)
+// Solidity: event CallFailed(bytes32 _opHash, uint256 _index, bytes _returnData)
 func (_WalletGuest *WalletGuestFilterer) FilterCallFailed(opts *bind.FilterOpts) (*WalletGuestCallFailedIterator, error) {
 
 	logs, sub, err := _WalletGuest.contract.FilterLogs(opts, "CallFailed")
@@ -444,9 +446,9 @@ func (_WalletGuest *WalletGuestFilterer) FilterCallFailed(opts *bind.FilterOpts)
 	return &WalletGuestCallFailedIterator{contract: _WalletGuest.contract, event: "CallFailed", logs: logs, sub: sub}, nil
 }
 
-// WatchCallFailed is a free log subscription operation binding the contract event 0x20832642214d5218c6428e71d8d2ddd9ad15a81ad2be8154d8c2e3ab08293fcb.
+// WatchCallFailed is a free log subscription operation binding the contract event 0x115f347c00e69f252cd6b63c4f81022a9564c6befe8aa719cb74640a4a306f0d.
 //
-// Solidity: event CallFailed(bytes32 _opHash, uint256 _index)
+// Solidity: event CallFailed(bytes32 _opHash, uint256 _index, bytes _returnData)
 func (_WalletGuest *WalletGuestFilterer) WatchCallFailed(opts *bind.WatchOpts, sink chan<- *WalletGuestCallFailed) (event.Subscription, error) {
 
 	logs, sub, err := _WalletGuest.contract.WatchLogs(opts, "CallFailed")
@@ -481,9 +483,9 @@ func (_WalletGuest *WalletGuestFilterer) WatchCallFailed(opts *bind.WatchOpts, s
 	}), nil
 }
 
-// ParseCallFailed is a log parse operation binding the contract event 0x20832642214d5218c6428e71d8d2ddd9ad15a81ad2be8154d8c2e3ab08293fcb.
+// ParseCallFailed is a log parse operation binding the contract event 0x115f347c00e69f252cd6b63c4f81022a9564c6befe8aa719cb74640a4a306f0d.
 //
-// Solidity: event CallFailed(bytes32 _opHash, uint256 _index)
+// Solidity: event CallFailed(bytes32 _opHash, uint256 _index, bytes _returnData)
 func (_WalletGuest *WalletGuestFilterer) ParseCallFailed(log types.Log) (*WalletGuestCallFailed, error) {
 	event := new(WalletGuestCallFailed)
 	if err := _WalletGuest.contract.UnpackLog(event, "CallFailed", log); err != nil {
@@ -628,9 +630,9 @@ func (_WalletGuest *WalletGuestFilterer) ParseCallSkipped(log types.Log) (*Walle
 	return event, nil
 }
 
-// WalletGuestCallSuccessIterator is returned from FilterCallSuccess and is used to iterate over the raw logs and unpacked data for CallSuccess events raised by the WalletGuest contract.
-type WalletGuestCallSuccessIterator struct {
-	Event *WalletGuestCallSuccess // Event containing the contract specifics and raw log
+// WalletGuestCallSucceededIterator is returned from FilterCallSucceeded and is used to iterate over the raw logs and unpacked data for CallSucceeded events raised by the WalletGuest contract.
+type WalletGuestCallSucceededIterator struct {
+	Event *WalletGuestCallSucceeded // Event containing the contract specifics and raw log
 
 	contract *bind.BoundContract // Generic contract to use for unpacking event data
 	event    string              // Event name to use for unpacking event data
@@ -644,7 +646,7 @@ type WalletGuestCallSuccessIterator struct {
 // Next advances the iterator to the subsequent event, returning whether there
 // are any more events found. In case of a retrieval or parsing error, false is
 // returned and Error() can be queried for the exact failure.
-func (it *WalletGuestCallSuccessIterator) Next() bool {
+func (it *WalletGuestCallSucceededIterator) Next() bool {
 	// If the iterator failed, stop iterating
 	if it.fail != nil {
 		return false
@@ -653,7 +655,7 @@ func (it *WalletGuestCallSuccessIterator) Next() bool {
 	if it.done {
 		select {
 		case log := <-it.logs:
-			it.Event = new(WalletGuestCallSuccess)
+			it.Event = new(WalletGuestCallSucceeded)
 			if err := it.contract.UnpackLog(it.Event, it.event, log); err != nil {
 				it.fail = err
 				return false
@@ -668,7 +670,7 @@ func (it *WalletGuestCallSuccessIterator) Next() bool {
 	// Iterator still in progress, wait for either a data or an error event
 	select {
 	case log := <-it.logs:
-		it.Event = new(WalletGuestCallSuccess)
+		it.Event = new(WalletGuestCallSucceeded)
 		if err := it.contract.UnpackLog(it.Event, it.event, log); err != nil {
 			it.fail = err
 			return false
@@ -684,42 +686,42 @@ func (it *WalletGuestCallSuccessIterator) Next() bool {
 }
 
 // Error returns any retrieval or parsing error occurred during filtering.
-func (it *WalletGuestCallSuccessIterator) Error() error {
+func (it *WalletGuestCallSucceededIterator) Error() error {
 	return it.fail
 }
 
 // Close terminates the iteration process, releasing any pending underlying
 // resources.
-func (it *WalletGuestCallSuccessIterator) Close() error {
+func (it *WalletGuestCallSucceededIterator) Close() error {
 	it.sub.Unsubscribe()
 	return nil
 }
 
-// WalletGuestCallSuccess represents a CallSuccess event raised by the WalletGuest contract.
-type WalletGuestCallSuccess struct {
+// WalletGuestCallSucceeded represents a CallSucceeded event raised by the WalletGuest contract.
+type WalletGuestCallSucceeded struct {
 	OpHash [32]byte
 	Index  *big.Int
 	Raw    types.Log // Blockchain specific contextual infos
 }
 
-// FilterCallSuccess is a free log retrieval operation binding the contract event 0xec670aed5ee1e72eb3eb601271be4b3f312e71f17eebdf10c1a0ab5a3af30ffd.
+// FilterCallSucceeded is a free log retrieval operation binding the contract event 0x5a589b1d8062f33451d29cae3dabd9b2e36c62aee644178c600977ca8dda661a.
 //
-// Solidity: event CallSuccess(bytes32 _opHash, uint256 _index)
-func (_WalletGuest *WalletGuestFilterer) FilterCallSuccess(opts *bind.FilterOpts) (*WalletGuestCallSuccessIterator, error) {
+// Solidity: event CallSucceeded(bytes32 _opHash, uint256 _index)
+func (_WalletGuest *WalletGuestFilterer) FilterCallSucceeded(opts *bind.FilterOpts) (*WalletGuestCallSucceededIterator, error) {
 
-	logs, sub, err := _WalletGuest.contract.FilterLogs(opts, "CallSuccess")
+	logs, sub, err := _WalletGuest.contract.FilterLogs(opts, "CallSucceeded")
 	if err != nil {
 		return nil, err
 	}
-	return &WalletGuestCallSuccessIterator{contract: _WalletGuest.contract, event: "CallSuccess", logs: logs, sub: sub}, nil
+	return &WalletGuestCallSucceededIterator{contract: _WalletGuest.contract, event: "CallSucceeded", logs: logs, sub: sub}, nil
 }
 
-// WatchCallSuccess is a free log subscription operation binding the contract event 0xec670aed5ee1e72eb3eb601271be4b3f312e71f17eebdf10c1a0ab5a3af30ffd.
+// WatchCallSucceeded is a free log subscription operation binding the contract event 0x5a589b1d8062f33451d29cae3dabd9b2e36c62aee644178c600977ca8dda661a.
 //
-// Solidity: event CallSuccess(bytes32 _opHash, uint256 _index)
-func (_WalletGuest *WalletGuestFilterer) WatchCallSuccess(opts *bind.WatchOpts, sink chan<- *WalletGuestCallSuccess) (event.Subscription, error) {
+// Solidity: event CallSucceeded(bytes32 _opHash, uint256 _index)
+func (_WalletGuest *WalletGuestFilterer) WatchCallSucceeded(opts *bind.WatchOpts, sink chan<- *WalletGuestCallSucceeded) (event.Subscription, error) {
 
-	logs, sub, err := _WalletGuest.contract.WatchLogs(opts, "CallSuccess")
+	logs, sub, err := _WalletGuest.contract.WatchLogs(opts, "CallSucceeded")
 	if err != nil {
 		return nil, err
 	}
@@ -729,8 +731,8 @@ func (_WalletGuest *WalletGuestFilterer) WatchCallSuccess(opts *bind.WatchOpts, 
 			select {
 			case log := <-logs:
 				// New log arrived, parse the event and forward to the user
-				event := new(WalletGuestCallSuccess)
-				if err := _WalletGuest.contract.UnpackLog(event, "CallSuccess", log); err != nil {
+				event := new(WalletGuestCallSucceeded)
+				if err := _WalletGuest.contract.UnpackLog(event, "CallSucceeded", log); err != nil {
 					return err
 				}
 				event.Raw = log
@@ -751,12 +753,12 @@ func (_WalletGuest *WalletGuestFilterer) WatchCallSuccess(opts *bind.WatchOpts, 
 	}), nil
 }
 
-// ParseCallSuccess is a log parse operation binding the contract event 0xec670aed5ee1e72eb3eb601271be4b3f312e71f17eebdf10c1a0ab5a3af30ffd.
+// ParseCallSucceeded is a log parse operation binding the contract event 0x5a589b1d8062f33451d29cae3dabd9b2e36c62aee644178c600977ca8dda661a.
 //
-// Solidity: event CallSuccess(bytes32 _opHash, uint256 _index)
-func (_WalletGuest *WalletGuestFilterer) ParseCallSuccess(log types.Log) (*WalletGuestCallSuccess, error) {
-	event := new(WalletGuestCallSuccess)
-	if err := _WalletGuest.contract.UnpackLog(event, "CallSuccess", log); err != nil {
+// Solidity: event CallSucceeded(bytes32 _opHash, uint256 _index)
+func (_WalletGuest *WalletGuestFilterer) ParseCallSucceeded(log types.Log) (*WalletGuestCallSucceeded, error) {
+	event := new(WalletGuestCallSucceeded)
+	if err := _WalletGuest.contract.UnpackLog(event, "CallSucceeded", log); err != nil {
 		return nil, err
 	}
 	event.Raw = log
