@@ -969,7 +969,7 @@ func TestTransactionToGuestModuleDeployAndCall(t *testing.T) {
 		}
 
 		guestAddress := testChain.V3SequenceContext().GuestModuleAddress
-		execdata := v3.ConstructCallsPayload(guestAddress, testChain.ChainID(), guestBundle, nil, nil).Encode(guestAddress)
+		execdata := v3.NewCallsPayload(guestAddress, testChain.ChainID(), guestBundle, nil, nil).Encode(guestAddress)
 
 		// Relay the txn manually, directly to the guest module
 		sender := testChain.GetRelayerWallet()
@@ -997,7 +997,8 @@ func TestTransactionToGuestModuleDeployAndCall(t *testing.T) {
 		assert.Equal(t, "2255", ret[0])
 
 		// Assert sequence.WaitForMetaTxn is able to find the metaTxnID
-		result, _, _, err := sequence.FetchMetaTransactionReceipt(context.Background(), testChain.ReceiptsListener, sequence.MetaTxnID(payload.Digest().String()[2:]))
+		metaTxnID, _, _ := sequence.ComputeMetaTxnIDFromCallsPayload(&payload)
+		result, _, _, err := sequence.FetchMetaTransactionReceipt(context.Background(), testChain.ReceiptsListener, metaTxnID)
 		assert.NoError(t, err)
 		assert.True(t, result.Status == sequence.MetaTxnExecuted)
 
@@ -1041,15 +1042,17 @@ func TestNonceOverflowChecks(t *testing.T) {
 	assert.Error(t, err)
 }
 
-func TestRawTransactionsEncoding(t *testing.T) {
-	data, err := hex.DecodeString("000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000016000000000000000000000000000000000000000000000000000000000000002a000000000000000000000000000000000000000000000000000000000000003e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000d4c4000000000000000000000000831de831a64405af965c67d6e0de2f9876fa2d9900000000000000000000000000000000000000000000000000005af3107a400000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c33f000000000000000000000000a804304f8e7c9ab5932ac3a4cc1468e40183d199000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000044a9059cbb000000000000000000000000831de831a64405af965c67d6e0de2f9876fa2d990000000000000000000000000000000000000000000000000000b5e620f480000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c33f000000000000000000000000be41c134fc3517cb0ec94b6eeafb66cf9998782f000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000044a9059cbb000000000000000000000000831de831a64405af965c67d6e0de2f9876fa2d99000000000000000000000000000000000000000000000000000110d9316ec0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c33f0000000000000000000000009400294c78f7dd244f4b5af6dfff6c971d25ad86000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000044a9059cbb000000000000000000000000831de831a64405af965c67d6e0de2f9876fa2d9900000000000000000000000000000000000000000000000000016bcc41e9000000000000000000000000000000000000000000000000000000000000")
+func TestTxnDecodeExecdata(t *testing.T) {
+	input := "0x1f6a1eb90000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000012000000000000000000000000000000000000000000000000000000000000000be13150c8f93229dc06f131943f44f82cfbbd53fcc26c5c9000084ad387c8a00000000000000000000000000000000000000000000000000000000000000370000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000311225500000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010c9c0000000000000000000000000000000000000000000000000000000000000000004404000171bab6d961b2b76c5d76b7fd92b962c3f0a1833098609f9b2444ead59af7507a9c31019b0f38998048658e88f3cd215cb9400c15509655655ec05257a11068544a00000000000000000000000000000000000000000000000000000000"
+
+	txns, nonce, _, err := sequence.DecodeExecdata(common.FromHex(input), common.HexToAddress("0x60349b199C2fEF7C3e1b53785dE5189493839b75"), big.NewInt(1337))
 	assert.NoError(t, err)
 
-	transactions, err := sequence.DecodeRawTransactions(data)
-	assert.NoError(t, err)
-	assert.Len(t, transactions, 4)
+	assert.Equal(t, txns[0].To, common.HexToAddress("0x8f93229DC06f131943F44F82CFbbD53FCC26C5C9"))
+	assert.Equal(t, txns[0].Value, big.NewInt(0))
+	assert.Equal(t, txns[0].GasLimit, big.NewInt(0))
+	assert.Equal(t, txns[0].DelegateCall, false)
+	assert.Equal(t, txns[0].RevertOnError, false)
 
-	data2, err := transactions.EncodeRaw()
-	assert.NoError(t, err)
-	assert.True(t, bytes.Equal(data, data2))
+	assert.Equal(t, nonce, big.NewInt(21))
 }
